Chore: 파일 구조 생성 및 프로젝트명 카멜케이스로 변경
