+++ app/App.tsx
@@ -1,0 +1,30 @@
+/**
+ * Sample React Native App
+ * https://github.com/facebook/react-native
+ *
+ * @format
+ */
+
+import React from "react";
+import { SafeAreaView, ScrollView, StatusBar, StyleSheet, useColorScheme } from "react-native";
+
+import { Colors } from "react-native/Libraries/NewAppScreen";
+
+function App(): React.JSX.Element {
+  const isDarkMode = useColorScheme() === "dark";
+
+  const backgroundStyle = {
+    backgroundColor: isDarkMode ? Colors.darker : Colors.lighter,
+  };
+
+  return (
+    <SafeAreaView style={backgroundStyle}>
+      <StatusBar barStyle={isDarkMode ? "light-content" : "dark-content"} backgroundColor={backgroundStyle.backgroundColor} />
+      <ScrollView contentInsetAdjustmentBehavior="automatic" style={backgroundStyle}></ScrollView>
+    </SafeAreaView>
+  );
+}
+
+const styles = StyleSheet.create({});
+
+export default App;
