+++ app/assets/fonts/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "fonts",
+  "version": "0.0.1"
+}
