+++ app/assets/images/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "images",
+  "version": "0.0.1"
+}
