+++ app/assets/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "assets",
+  "version": "0.0.1"
+}
