+++ app/components/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "components",
+  "version": "0.0.1"
+}
