+++ app/screens/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "screens",
+  "version": "0.0.1"
+}
