+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "Michi",
+  "name": "michi",
   "version": "0.0.1",
   "private": true,
   "scripts": {
